requirements.txt
beautifulsoup4
Flask
Flask-Cors
mwapi
mwparserfromhell
protobuf
PyYAML
requests
sentence-transformers